requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
torch>=2.0.0
numpy>=1.26.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Document parsing
python-docx>=1.1.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
